+++ src/v3/plugins/git/example/example-git.json
