+++ src/v3/plugins/git/example/synchronizeToGithub.sh
