+++ src/v3/plugins/github/example/example-github.json
